content_moderation_webScraper/requirements.txt
boto3~=1.39.3
Flask~=3.1.1
botocore~=1.39.3
requests~=2.31.0
beautifulsoup4~=4.12.2
Pillow~=10.0.0
pytest~=7.4.0
